Rename OpenInColab.tsx => index.tsx
